rework to api

+++ BMelt.ClassLibrary/Repository/RecipeRepository.cs
@@ -1,10 +1,10 @@
-﻿using BMelt.ClassLibrary.Enums;
-using BMelt.ClassLibrary.Models;
+﻿using BMelt.ClassLibrary.Models;
+using BMelt.ClassLibrary.Repository.Interface;
 using Microsoft.EntityFrameworkCore;
 
 namespace BMelt.ClassLibrary.Repository
 {
-    public class RecipeRepository
+    public class RecipeRepository : IRecipeRepository
     {
         private readonly DatabaseContext _dbContext;
 
@@ -15,7 +15,7 @@ namespace BMelt.ClassLibrary.Repository
 
         public async Task<Recipe> GetRecipeAsync(Guid id)
         {
-            return await _dbContext.Recipes.FindAsync(id);
+            return await _dbContext.Recipes.FindAsync(id) ?? new Recipe();
         }
 
         public async Task<IEnumerable<Recipe>> GetRecipesAsync(Cuisine cuisine)
@@ -47,10 +47,16 @@ namespace BMelt.ClassLibrary.Repository
             return recipe;
         }
 
-        public async Task DeleteRecipeAsync(Recipe recipe)
+        public async Task<bool> DeleteRecipeAsync(Guid id)
         {
-            _dbContext.Recipes.Remove(recipe);
-            await _dbContext.SaveChangesAsync();
+            var productExist = _dbContext.Recipes.Find(recipe.Id);
+            if (productExist != null)
+            {
+                _dbContext.Recipes.Remove(recipe);
+                await _dbContext.SaveChangesAsync();
+                return true;
+            }
+            return false;
         }
     }
 }
